feat: add model configuration logging, force bfloat16 for FlashAttention, disable torch.compile, and update DataLoader to ignore SIGINT

+++ src/model/swa.py
@@ -80,7 +80,7 @@ class SlidingWindowAttention(nn.Module):
         # Flash Attention 2 with sliding window
         # flash_attn_func expects (B, T, H, D) layout — already correct
         attn_output = flash_attn_func(
-            q, k, v,
+            q.to(torch.bfloat16), k.to(torch.bfloat16), v.to(torch.bfloat16),
             causal=True,
             window_size=(self.window_size, 0),  # (left_window, right_window=0 for causal)
         )
